Be careful of overflows and int types
- 60*1000 should be 60000, but that's more than INT_MAX

+++ DanceWithLanceArbduino_w_Sign/DanceWithLanceArbduino_w_Sign.ino
@@ -79,7 +79,7 @@ const int ledPins[NUM_LEDS] = {
 
 // long-term memory in EEPROM
 #define WATTHOURS_EEPROM_ADDRESS 20
-#define BACKUP_INTERVAL 60*1000
+#define BACKUP_INTERVAL ((long)60*1000)
 
 #define RESET_PIN 12
 #define BUCK_ENABLE_PIN 11
